#13 - 장바구니 기능 로직 구현 #13 내용 바탕으로 기능 구현 진행 Order Page, Detail Page 오류 발생 및 수정 진행 ReadMe 일부 변경 진행

+++ src/main/java/org/practice/shop/service/ItemImgServiceImpl.java
@@ -53,8 +53,10 @@ public class ItemImgServiceImpl implements ItemImgService{
         // itemImgUpdate
         Optional<ItemImg> result = itemImgRepository.findById(itemImgId);
         ItemImg itemImg = result.get();
+        // 원본 이미지 보관
+        String oriImgName = multipartFile.getOriginalFilename();
         // 이미지 이름이 비어있지 않으면
-        if (!StringUtils.isEmpty(itemImg.getImgName())) {
+        if (!StringUtils.isEmpty(itemImg.getImgName()) && oriImgName.length() > 0) {
             fileService.deleteFile(itemImgLocation + "/" + itemImg.getImgName());
         }
         // 이미지 업로드
@@ -62,7 +64,7 @@ public class ItemImgServiceImpl implements ItemImgService{
         if(!StringUtils.isEmpty(oriImgName1)) {
             String imgName = fileService.uploadFile(itemImgLocation, oriImgName1, multipartFile.getBytes());
             // 상품 이미지
-            String imgUrl = "/images/item" + imgName;
+            String imgUrl = "/images/item/" + imgName;
             itemImg.update(oriImgName1, imgName, imgUrl);
             itemImgRepository.save(itemImg);
         }
